Fix 'init' command for logged-out gcloud session
The 'init' command would hang because it would fail to get bucket via gcloud if
session was logged out. To fix this I moved the variable to the script
after the user is logged it, instead of having it as a make variable.

Also prepend shell variables with underscore to easily distinguish them
from make variables.

Added a few small refactorings here and there, and renamed 'prep'
command to 'init'.